// src/app/interfaces/form-abilities-values.ts
export interface FormAbilitiesValues {
  abilitiesProgression: {
    l1: string;
    l2: string;
    l3: string;
    l4: string;
    l5: string;
    l6: string;
    l7: string;
    l8: string;
    l9: string;
    l10: string;
    l11: string;
    l12: string;
    l13: string;
    l14: string;
    l15: string;
    l16: string;
    l17: string;
    l18: string;
  };
  abilitiesProgressionDescription: string;
}